Remove data providers search
