--- components/repositories-browser/repositories-browser.module.scss
@@ -1,9 +1,0 @@
-@import "components/bootstrap/core";
-
-.data-providers-card {
-  transition: background-color 0.25s;
-  &:hover,
-  &:focus-within {
-    background-color: $gray-100;
-  }
-}
