+++ package.json
@@ -38,7 +38,6 @@
     "cssnano": "^4.1.10",
     "decko": "^1.2.0",
     "formdata-polyfill": "^3.0.20",
-    "fuse.js": "^3.6.1",
     "js-cookie": "^2.2.1",
     "json-loader": "^0.5.7",
     "leaflet": "^1.7.1",
